+++ app/src/main/scala/serv1/model/HistoricalData.scala
@@ -1,10 +1,10 @@
 package serv1.model
 
 case class HistoricalData(
-                      var timestamp: Long,
-                      var high: Long,
-                      var low: Long,
-                      var open: Long,
-                      var close: Long,
-                      var vol: Long
-                    )
+                           var timestamp: Long,
+                           var high: Long,
+                           var low: Long,
+                           var open: Long,
+                           var close: Long,
+                           var vol: Double
+                         )
